requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
Pillow
celery
redis
drf-spectacular
drf-spectacular[sidecar]
gunicorn
